+++ tests/robots/test_robot.py
@@ -1,3 +1,4 @@
+import math
 import os
 import re
 
@@ -11,8 +12,8 @@ from compas_fab.backends.interfaces import ClientInterface
 from compas_fab.backends.interfaces import InverseKinematics
 from compas_fab.backends.interfaces import PlannerInterface
 from compas_fab.robots import Robot
-from compas_fab.robots import Tool
 from compas_fab.robots import RobotSemantics
+from compas_fab.robots import Tool
 from compas_fab.robots.ur5 import Robot as Ur5Robot
 
 BASE_FOLDER = os.path.dirname(__file__)
@@ -20,12 +21,12 @@ BASE_FOLDER = os.path.dirname(__file__)
 
 @pytest.fixture
 def panda_srdf():
-    return os.path.join(BASE_FOLDER, 'fixtures', 'panda_semantics.srdf')
+    return os.path.join(BASE_FOLDER, "fixtures", "panda_semantics.srdf")
 
 
 @pytest.fixture
 def panda_urdf():
-    return os.path.join(BASE_FOLDER, 'fixtures', 'panda.urdf')
+    return os.path.join(BASE_FOLDER, "fixtures", "panda.urdf")
 
 
 @pytest.fixture
@@ -62,8 +63,16 @@ def ur5_with_fake_ik(ur5_robot_instance, fake_client):
     class FakeInverseKinematics(InverseKinematics):
         def inverse_kinematics(self, robot, frame_WCF, start_configuration=None, group=None, options=None):
             results = [
-                ((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint')),
-                ((-2.238, -3.175, 2.174, 4.143, -5.616, -6.283), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint')),
+                (
+                    (-1.572, -2.560, 2.196, 2.365, 0.001, 1.137),
+                    (0, 0, 0, 0, 0, 0),
+                    ("shoulder_pan_joint", "shoulder_lift_joint", "elbow_joint", "wrist_1_joint", "wrist_2_joint", "wrist_3_joint"),
+                ),
+                (
+                    (-2.238, -3.175, 2.174, 4.143, -5.616, -6.283),
+                    (0, 0, 0, 0, 0, 0),
+                    ("shoulder_pan_joint", "shoulder_lift_joint", "elbow_joint", "wrist_1_joint", "wrist_2_joint", "wrist_3_joint"),
+                ),
             ]
 
             for ik in results:
@@ -101,158 +110,158 @@ def fake_client():
 
 @pytest.fixture
 def robot_tool1():
-    mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
+    mesh = Mesh.from_stl(compas_fab.get("planning_scene/cone.stl"))
     frame = Frame([0.14, 0, 0], [0, 1, 0], [0, 0, 1])
     return Tool(mesh, frame)
 
 
 @pytest.fixture
 def robot_tool2():
-    mesh = Mesh.from_stl(compas_fab.get('planning_scene/cone.stl'))
+    mesh = Mesh.from_stl(compas_fab.get("planning_scene/cone.stl"))
     frame = Frame([0.89, 0, 0], [0, 1, 0], [0, 0, 1])
     return Tool(mesh, frame)
 
 
 def test_basic_name_only():
-    robot = Robot.basic('testbot')
+    robot = Robot.basic("testbot")
     assert robot.artist is None
 
 
 def test_basic_name_joints_links(ur5_joints, ur5_links):
-    robot = Robot.basic('testbot', joints=ur5_joints, links=ur5_links)
+    robot = Robot.basic("testbot", joints=ur5_joints, links=ur5_links)
     assert len(robot.model.links) == 11
     assert len(robot.get_configurable_joint_names()) == 6
 
 
 def test_basic_name_joints(ur5_joints):
-    robot = Robot.basic('testbot', joints=ur5_joints)
+    robot = Robot.basic("testbot", joints=ur5_joints)
     assert len(robot.model.joints) == 10
 
 
 def test_basic_name_links(ur5_links):
-    robot = Robot.basic('testbot', links=ur5_links)
+    robot = Robot.basic("testbot", links=ur5_links)
     assert len(robot.model.links) == 11
 
 
 def test_basic_attr():
-    robot = Robot.basic('testbot', location='rfl')
-    assert robot.model.attr['location'] == 'rfl'
+    robot = Robot.basic("testbot", location="rfl")
+    assert robot.model.attr["location"] == "rfl"
 
 
 def test_name(panda_robot_instance):
     robot = panda_robot_instance
-    assert robot.name == 'panda'
+    assert robot.name == "panda"
 
 
 def test_group_names(ur5_robot_instance):
     robot = ur5_robot_instance
-    assert sorted(robot.group_names) == ['endeffector', 'manipulator']
+    assert sorted(robot.group_names) == ["endeffector", "manipulator"]
 
 
 def test_main_group_name(panda_robot_instance):
     robot = panda_robot_instance
-    assert robot.main_group_name == 'panda_arm_hand'
+    assert robot.main_group_name == "panda_arm_hand"
 
 
 def test_root_name(ur5_robot_instance):
     robot = ur5_robot_instance
-    assert robot.root_name == 'world'
+    assert robot.root_name == "world"
 
 
 def test_get_end_effector_link_name(panda_robot_instance):
     robot = panda_robot_instance
-    assert robot.get_end_effector_link_name(group=None) == 'panda_rightfinger'
-    assert robot.get_end_effector_link_name(group='panda_arm_hand') == 'panda_rightfinger'
-    assert robot.get_end_effector_link_name(group='panda_arm') == 'panda_link8'
+    assert robot.get_end_effector_link_name(group=None) == "panda_rightfinger"
+    assert robot.get_end_effector_link_name(group="panda_arm_hand") == "panda_rightfinger"
+    assert robot.get_end_effector_link_name(group="panda_arm") == "panda_link8"
 
 
 def test_get_end_effector_link_name_wrong_group(panda_robot_instance):
     robot = panda_robot_instance
     with pytest.raises(KeyError):
-        robot.get_end_effector_link_name(group='panda_leg')
+        robot.get_end_effector_link_name(group="panda_leg")
 
 
 def test_get_end_effector_link(ur5_robot_instance):
     robot = ur5_robot_instance
 
     assert robot.get_end_effector_link(group=None).name == "ee_link"
-    assert robot.get_end_effector_link(group='endeffector').name == "ee_link"
+    assert robot.get_end_effector_link(group="endeffector").name == "ee_link"
 
 
 def test_get_end_effector_frame(panda_robot_instance):
     robot = panda_robot_instance
-    assert round(robot.get_end_effector_frame(group=None).point.x, 3) == .301
-    assert round(robot.get_end_effector_frame(group='panda_arm').point.x, 3) == .359
+    assert round(robot.get_end_effector_frame(group=None).point.x, 3) == 0.301
+    assert round(robot.get_end_effector_frame(group="panda_arm").point.x, 3) == 0.359
 
 
 def test_get_base_link_name(ur5_robot_instance):
     robot = ur5_robot_instance
 
-    assert robot.get_base_link_name(group=None) == 'base_link'
-    assert robot.get_base_link_name(group='endeffector') == 'ee_link'
+    assert robot.get_base_link_name(group=None) == "base_link"
+    assert robot.get_base_link_name(group="endeffector") == "ee_link"
 
 
 def test_get_base_link_name_wo_semantics(panda_robot_instance_wo_semantics):
     robot = panda_robot_instance_wo_semantics
 
-    assert robot.get_base_link_name(group=None) == 'panda_link0'
-    assert robot.get_base_link_name(group='endeffector') == 'panda_link0'
+    assert robot.get_base_link_name(group=None) == "panda_link0"
+    assert robot.get_base_link_name(group="endeffector") == "panda_link0"
 
 
 def test_get_base_link(panda_robot_instance):
     robot = panda_robot_instance
 
     link = robot.get_base_link(group=None)
-    assert link.name == 'panda_link0'
+    assert link.name == "panda_link0"
 
-    link = robot.get_base_link(group='panda_arm')
-    assert link.name == 'panda_link0'
+    link = robot.get_base_link(group="panda_arm")
+    assert link.name == "panda_link0"
 
-    link = robot.get_base_link(group='hand')
-    assert link.name == 'panda_hand'
+    link = robot.get_base_link(group="hand")
+    assert link.name == "panda_hand"
 
 
 def test_get_base_link_wo_semantics(panda_robot_instance_wo_semantics):
     robot = panda_robot_instance_wo_semantics
 
     link = robot.get_base_link(group=None)
-    assert link.name == 'panda_link0'
+    assert link.name == "panda_link0"
 
-    link = robot.get_base_link(group='panda_arm')
-    assert link.name == 'panda_link0'
+    link = robot.get_base_link(group="panda_arm")
+    assert link.name == "panda_link0"
 
-    link = robot.get_base_link(group='hand')
-    assert link.name == 'panda_link0'
+    link = robot.get_base_link(group="hand")
+    assert link.name == "panda_link0"
 
 
 def test_get_base_frame(panda_robot_instance):
     robot = panda_robot_instance
 
     assert robot.get_base_frame(group=None) == Frame.worldXY()
-    assert robot.get_base_frame(group='panda_arm') == Frame.worldXY()
+    assert robot.get_base_frame(group="panda_arm") == Frame.worldXY()
 
 
 def test_get_base_frame_w_artist(panda_robot_instance_w_artist):
     robot = panda_robot_instance_w_artist
 
     assert robot.get_base_frame(group=None) == Frame.worldXY()
-    assert robot.get_base_frame(group='panda_arm') == Frame.worldXY()
+    assert robot.get_base_frame(group="panda_arm") == Frame.worldXY()
 
 
 def test_get_base_frame_when_link_has_parent(ur5_robot_instance):
     robot = ur5_robot_instance
-    base_frame = robot.get_base_frame(group='endeffector')
+    base_frame = robot.get_base_frame(group="endeffector")
 
     assert [round(v, 3) for v in list(base_frame.point)] == [0.817, 0.191, -0.005]
-    assert [round(v) for v in base_frame.data['xaxis']] == [0, 1, 0]
-    assert [round(v) for v in base_frame.data['yaxis']] == [1, 0, 0]
+    assert [round(v) for v in base_frame.data["xaxis"]] == [0, 1, 0]
+    assert [round(v) for v in base_frame.data["yaxis"]] == [1, 0, 0]
 
 
 def test_get_configurable_joints(ur5_robot_instance):
     robot = ur5_robot_instance
     joints = robot.get_configurable_joints()
 
-    pattern = re.compile(r'(shoulder|wrist|elbow).*_joint')
+    pattern = re.compile(r"(shoulder|wrist|elbow).*_joint")
     matches = [pattern.match(joint.name) for joint in joints]
     assert all(matches)
 
@@ -261,7 +270,7 @@ def test_get_configurable_joints_wo_semantics(panda_robot_instance_wo_semantics)
     robot = panda_robot_instance_wo_semantics
     joints = robot.get_configurable_joints()
 
-    pattern = re.compile(r'panda_.*joint\d')
+    pattern = re.compile(r"panda_.*joint\d")
     matches = [pattern.match(joint.name) for joint in joints]
     assert all(matches)
 
@@ -273,11 +282,20 @@ def test_inverse_kinematics_repeated_calls_will_return_next_result(ur5_with_fake
     start_config = robot.zero_configuration()
 
     configuration = robot.inverse_kinematics(frame, start_config)
-    assert str(configuration) == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    assert (
+        str(configuration)
+        == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    )
     configuration = robot.inverse_kinematics(frame, start_config)
-    assert str(configuration) == "Configuration((-2.238, -3.175, 2.174, 4.143, -5.616, -6.283), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    assert (
+        str(configuration)
+        == "Configuration((-2.238, -3.175, 2.174, 4.143, -5.616, -6.283), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    )
     configuration = robot.inverse_kinematics(frame, start_config)
-    assert str(configuration) == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    assert (
+        str(configuration)
+        == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    )
 
 
 def test_iter_inverse_kinematics(ur5_with_fake_ik):
@@ -288,8 +306,31 @@ def test_iter_inverse_kinematics(ur5_with_fake_ik):
 
     solutions = list(robot.iter_inverse_kinematics(frame, start_config))
     assert len(solutions) == 2
-    assert str(solutions[0]) == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
-    assert str(solutions[1]) == "Configuration((-2.238, -3.175, 2.174, 4.143, -5.616, -6.283), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    assert (
+        str(solutions[0])
+        == "Configuration((-1.572, -2.560, 2.196, 2.365, 0.001, 1.137), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    )
+    assert (
+        str(solutions[1])
+        == "Configuration((-2.238, -3.175, 2.174, 4.143, -5.616, -6.283), (0, 0, 0, 0, 0, 0), ('shoulder_pan_joint', 'shoulder_lift_joint', 'elbow_joint', 'wrist_1_joint', 'wrist_2_joint', 'wrist_3_joint'))"
+    )
+
+
+def test_forward_kinematics_without_tool(ur5_robot_instance):
+    robot = ur5_robot_instance
+
+    frame_t0cf = robot.forward_kinematics(robot.zero_configuration())
+    assert str(frame_t0cf) == "Frame(Point(0.817, 0.191, -0.005), Vector(-0.000, 1.000, 0.000), Vector(1.000, 0.000, 0.000))"
+
+
+def test_forward_kinematics_with_tool(ur5_robot_instance, robot_tool1):
+    robot = ur5_robot_instance
+    tool = robot_tool1
+
+    robot.attach_tool(tool)
+    frame_t0cf = robot.forward_kinematics(robot.zero_configuration(), use_attached_tool_frame=False)
+    frame_tcf = robot.forward_kinematics(robot.zero_configuration(), use_attached_tool_frame=True)
+    assert math.fabs(frame_t0cf.point.y - frame_tcf.point.y) == tool.frame.point.x
 
 
 def test_attach_tool_without_group(ur5_robot_instance, robot_tool1):
@@ -361,7 +402,7 @@ def test_detach_tool_with_group(ur5_robot_instance, robot_tool1):
     assert len(robot.attached_tools) == 0
 
 
-def test_wrond_gorup_name_raises_exception(ur5_robot_instance, robot_tool1):
+def test_wrong_group_name_raises_exception(ur5_robot_instance, robot_tool1):
     robot = ur5_robot_instance
     tool = robot_tool1
     wrong_group_name = "theavengers"